Renamed files and modified readme to reflect package structure.
