+++ .gitignore
@@ -1,3 +1,6 @@
+#config
+.json
+
 #contains ignore for python and React/JS stuff
 __pycache__/
 
@@ -25,4 +28,3 @@ yarn-error.log*
 
 # don't want VSCode stuff stored in repo
 .vscode
-settings.json
