+++ tic-react/src/js/requests.js
@@ -1,3 +1,3 @@
 /*Constants for making requests will be stored here, so can be used by fetch
 in other files. */
-export const BASE_URL = "http://127.0.0.1:5000/";
+export const BASE_URL = "localhost:5000";
